+++ lib/mir/src/analysis/lambda_lifting/escape.rs
